# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Aqar Real Estate Platform API
 * OpenAPI spec version: 0.1.0
 */
export type HealthStatusStatus =
  (typeof HealthStatusStatus)[keyof typeof HealthStatusStatus];

export const HealthStatusStatus = {
  ok: "ok",
  error: "error",
} as const;

export type HealthStatusDb =
  (typeof HealthStatusDb)[keyof typeof HealthStatusDb];

export const HealthStatusDb = {
  ok: "ok",
  error: "error",
} as const;

export interface HealthStatus {
  status: HealthStatusStatus;
  db?: HealthStatusDb;
  uptimeSeconds?: number;
  dbLatencyMs?: number;
  timestamp?: string;
}

export interface ErrorResponse {
  error: string;
}

export interface ErrorEnvelope {
  error: string;
}

export interface UploadUrlRequest {
  /** @minLength 1 */
  name: string;
  /** @minimum 1 */
  size: number;
  /** @minLength 1 */
  contentType: string;
}

export interface UploadUrlResponse {
  uploadURL: string;
  objectPath: string;
}

export interface Property {
  id: number;
  referenceId: string;
  title: string;
  titleAr: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  descriptionAr?: string | null;
  status: string;
  type: string;
  price: number;
  currency: string;
  /** @nullable */
  area?: number | null;
  /** @nullable */
  bedrooms?: number | null;
  /** @nullable */
  bathrooms?: number | null;
  /** @nullable */
  furnished?: string | null;
  featured: boolean;
  active: boolean;
  views: number;
  whatsappClicks: number;
  callClicks: number;
  /** @nullable */
  governorateId?: number | null;
  /** @nullable */
  governorateName?: string | null;
  /** @nullable */
  areaId?: number | null;
  /** @nullable */
  areaName?: string | null;
  /** @nullable */
  officeId?: number | null;
  /** @nullable */
  officeName?: string | null;
  /** @nullable */
  officeLogo?: string | null;
  /** @nullable */
  primaryImage?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface PropertyImage {
  id: number;
  url: string;
  isPrimary: boolean;
  sortOrder: number;
}

export interface Office {
  id: number;
  name: string;
  nameAr: string;
  slug: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  descriptionAr?: string | null;
  /** @nullable */
  logo?: string | null;
  /** @nullable */
  coverImage?: string | null;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  whatsapp?: string | null;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  website?: string | null;
  /** @nullable */
  instagram?: string | null;
  /** @nullable */
  twitter?: string | null;
  /** @nullable */
  governorateId?: number | null;
  /** @nullable */
  governorateName?: string | null;
  verified: boolean;
  featured: boolean;
  active: boolean;
  totalListings: number;
  activeListings: number;
  /** @nullable */
  planName?: string | null;
  createdAt: string;
}

export type PropertyDetail = Property & {
  images?: PropertyImage[];
  amenities?: string[];
  office?: Office;
};

export interface PropertyListResponse {
  properties: Property[];
  total: number;
  page: number;
  limit: number;
  totalPages: number;
}

export type OfficeDetail = Office & {
  totalViews?: number;
  totalLeads?: number;
};

export interface OfficeListResponse {
  offices: Office[];
  total: number;
  page: number;
  limit: number;
  totalPages: number;
}

export interface Lead {
  id: number;
  customerName: string;
  phone: string;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  message?: string | null;
  inquiryType: string;
  status: string;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  propertyId?: number | null;
  /** @nullable */
  propertyTitle?: string | null;
  /** @nullable */
  propertyRef?: string | null;
  /** @nullable */
  officeId?: number | null;
  /** @nullable */
  sourcePage?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface OfficeStats {
  totalListings: number;
  activeListings: number;
  featuredListings: number;
  totalViews: number;
  totalLeads: number;
  whatsappClicks: number;
  callClicks: number;
  recentLeads: Lead[];
}

export interface CreateLeadBody {
  customerName: string;
  phone: string;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  message?: string | null;
  inquiryType: string;
  /** @nullable */
  propertyId?: number | null;
  /** @nullable */
  officeId?: number | null;
  /** @nullable */
  sourcePage?: string | null;
}

export interface UpdateLeadBody {
  /** @nullable */
  status?: string | null;
  /** @nullable */
  notes?: string | null;
}

export interface Governorate {
  id: number;
  name: string;
  nameAr: string;
  countryCode: string;
}

export interface Area {
  id: number;
  name: string;
  nameAr: string;
  governorateId: number;
  /** @nullable */
  governorateName?: string | null;
}

export interface SubscriptionPlan {
  id: number;
  name: string;
  nameAr: string;
  price: number;
  currency: string;
  maxListings: number;
  featuredListings: number;
  hasLeadDashboard: boolean;
  hasAnalytics: boolean;
  hasWhatsappSupport: boolean;
  hasPriorityPlacement: boolean;
  hasCustomProfile: boolean;
  features: string[];
}

export interface PlatformStats {
  totalProperties: number;
  totalOffices: number;
  totalCities: number;
  totalLeads: number;
}

export interface DashboardStats {
  totalListings: number;
  activeListings: number;
  featuredListings: number;
  totalViews: number;
  totalLeads: number;
  newLeads: number;
  whatsappClicks: number;
  callClicks: number;
  recentLeads: Lead[];
  topProperties: Property[];
}

export type ListPropertiesParams = {
  /**
   * @nullable
   */
  status?: string | null;
  /**
   * @nullable
   */
  type?: string | null;
  /**
   * @nullable
   */
  governorateId?: number | null;
  /**
   * @nullable
   */
  areaId?: number | null;
  /**
   * @nullable
   */
  minPrice?: number | null;
  /**
   * @nullable
   */
  maxPrice?: number | null;
  /**
   * @nullable
   */
  bedrooms?: number | null;
  /**
   * @nullable
   */
  bathrooms?: number | null;
  /**
   * @nullable
   */
  furnished?: string | null;
  /**
   * @nullable
   */
  featured?: string | null;
  /**
   * @nullable
   */
  officeId?: number | null;
  /**
   * @nullable
   */
  keyword?: string | null;
  /**
   * @nullable
   */
  sort?: string | null;
  /**
   * @nullable
   */
  page?: number | null;
  /**
   * @nullable
   */
  limit?: number | null;
};

export type GetLatestPropertiesParams = {
  /**
   * @nullable
   */
  limit?: number | null;
};

export type ListOfficesParams = {
  /**
   * @nullable
   */
  governorateId?: number | null;
  /**
   * @nullable
   */
  verified?: string | null;
  /**
   * @nullable
   */
  page?: number | null;
  /**
   * @nullable
   */
  limit?: number | null;
};

export type GetOfficePropertiesParams = {
  /**
   * @nullable
   */
  page?: number | null;
  /**
   * @nullable
   */
  limit?: number | null;
};

export type ListLeadsParams = {
  /**
   * @nullable
   */
  officeId?: number | null;
  /**
   * @nullable
   */
  status?: string | null;
  /**
   * @nullable
   */
  propertyId?: number | null;
};

export type ListAreasParams = {
  /**
   * @nullable
   */
  governorateId?: number | null;
};