# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Aqar Real Estate Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /leads:
    get:
      operationId: listLeads
      tags: [leads]
      summary: List leads for an office
      parameters:
        - name: officeId
          in: query
          schema:
            type: ["integer", "null"]
        - name: status
          in: query
          schema:
            type: ["string", "null"]
        - name: propertyId
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: List of leads
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Lead"

    post:
      operationId: createLead
      tags: [leads]
      summary: Submit a property inquiry / lead
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateLeadBody"
      responses:
        "201":
          description: Lead created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lead"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /leads/{id}:
    patch:
      operationId: updateLead
      tags: [leads]
      summary: Update lead status or notes
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateLeadBody"
      responses:
        "200":
          description: Updated lead
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lead"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    Lead:
      type: object
      properties:
        id:
          type: integer
        customerName:
          type: string
        phone:
          type: string
        email:
          type: ["string", "null"]
        message:
          type: ["string", "null"]
        inquiryType:
          type: string
        status:
          type: string
        notes:
          type: ["string", "null"]
        propertyId:
          type: ["integer", "null"]
        propertyTitle:
          type: ["string", "null"]
        propertyRef:
          type: ["string", "null"]
        officeId:
          type: ["integer", "null"]
        sourcePage:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - customerName
        - phone
        - inquiryType
        - status
        - createdAt
        - updatedAt
    CreateLeadBody:
      type: object
      properties:
        customerName:
          type: string
        phone:
          type: string
        email:
          type: ["string", "null"]
        message:
          type: ["string", "null"]
        inquiryType:
          type: string
        propertyId:
          type: ["integer", "null"]
        officeId:
          type: ["integer", "null"]
        sourcePage:
          type: ["string", "null"]
      required:
        - customerName
        - phone
        - inquiryType
    UpdateLeadBody:
      type: object
      properties:
        status:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
...